pyproject.toml
[project]
name = "evoapi-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "evolutionapi>=0.1.1",
    "mcp[cli]>=1.6.0",
    "numpy>=2.3.0",
    "python-dotenv>=1.1.0",
]
